feat(client): entities diagram component
